requirements.txt
asgiref==3.3.4
astroid==2.5.6
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-autocomplete-light==3.8.2
django-ckeditor==6.1.0
django-filter==2.4.0
django-js-asset==1.2.2
django-querysetsequence==0.14
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pylint==2.8.3
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
wrapt==1.12.1
